requirements.txt
# Production dependencies
youtube-transcript-api==1.2.3
openai>=1.0.0
pytube>=15.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0